+++ react/components/TranslatedTitle.tsx
@@ -1,11 +1,9 @@
 import React, { FunctionComponent } from 'react'
-import { injectIntl, InjectedIntlProps } from 'react-intl'
+import { FormattedMessage } from 'react-intl'
 
 // The b-s token makes the `b` class be applied only on mobile devices.
-const TranslatedTile: FunctionComponent<
-  TranslatedTileProps & InjectedIntlProps
-> = ({ intl, title }) => {
-  const titleText = title || intl.formatMessage({ id: 'base.title' })
+const TranslatedTile: FunctionComponent<TranslatedTileProps> = ({ title }) => {
+  const titleText = title || <FormattedMessage id="base.title" />
   return (
     <div className="flex flex-grow-1 justify-center b-s">
       <span className="t-heading-4">{titleText}</span>
@@ -17,4 +15,4 @@ interface TranslatedTileProps {
   title?: string
 }
 
-export default injectIntl(TranslatedTile)
+export default TranslatedTile